requirements.txt
flask
flask_sqlalchemy
flask_wtf
wtforms
email_validator
flask_login
passlib[bcrypt]
python-dotenv
pytz
sqlalchemy-utils
pandas
bcrypt
